Add files via upload
The surveillance robot will be designed to deliver a reasonable level of efficiency and simplicity, providing each user with a streamlined user experience. The surveillance robot is aimed at providing monitoring inclusive of vision, motion, fire, and carbon monoxide with a limited setup. The surveillance robot can be customized to fuse seamlessly to any home, apartments or multi-dwelling units. Based on modular designs and complete scalability, the surveillance robot is designed to be expandable and allow for future home control upgrades, thus enhancing the protection of your home as time and lifestyles change.



FUTURE SCOPE
1. The most important use of this project will be for surveillance.
2. It can be used to infiltrate at various locations where it would be hard for a human to 
   infiltrate without getting noticed.
3.Can be used for household surveillance.
4.Can be used by traffic security to assess fast moving cars
5. A drone can be detected using radar system whereas it is hard for a land rover to be detected     
   while it is moving on land.
6.Can be used to hijacked areas/buildings.
7.Can be used under a tunnel with an introduction of sensors like gas and pollution sensor.